Crate efi_sys

Low-level UEFI definitions.

This crate provides the bare minimum to make Gensokyo work, and so is in no way near complete.

Comments in the code refer to the UEFI Specification 2.6, available at http://www.uefi.org/sites/default/files/resources/UEFI%20Spec%202_6.pdf.

Modules

graphics_output

simple_text_input

simple_text_output

Structs

BltPixel

BootServices

ConfigurationTable

Daylight

DebugDisposition

GraphicsOutputProtocol

Guid

MemoryAttribute

MemoryDescriptor
Represents a UEFI memory descriptor.

Mode

ModeInformation

OpenProtocolAttribute

OpenProtocolInformationEntry

PixelBitmask

RuntimeServices

SimpleTextInputProtocol

SimpleTextOutputProtocol

SystemTable

TableHeader

Time

TimeCapabilities

Enums

AllocateType

BltOperation

KnownStatus

LocateSearchType

MemoryType

PixelFormat

ResetType

Void

Constants

ADJUST_DAYLIGHT

BY_CHILD_CONTROLLER

BY_DRIVER

BY_HANDLE_PROTOCOL

EXCLUSIVE

GET_PROTOCOL

GRAPHICS_OUTPUT_PROTOCOL_GUID

IN_DAYLIGHT

MAX_BIT

MEMORY_DESCRIPTOR_VERSION

MEMORY_MORE_RELIABLE

MEMORY_NV

MEMORY_RO

MEMORY_RP

MEMORY_RUNTIME

MEMORY_UC

MEMORY_UCE

MEMORY_WB

MEMORY_WC

MEMORY_WP

MEMORY_WT

MEMORY_XP

OPTIONAL_PTR

PAGE_SIZE

SIMPLE_TEXT_OUTPUT_GUID

TEST_PROTOCOL

UNSPECIFIED_TIMEZONE

Type Definitions

AllocatePages

AllocatePool

Blt

CloseProtocol

ConvertPointer

Exit

ExitBootServices

FreePages

FreePool

GetMemoryMap

GetNextMonotonicCount

GetTime

GetWakeupTime

Handle

LocateHandleBuffer

LocateProtocol

OpenProtocol

OpenProtocolInformation

PhysicalAddress

ProtocolsPerHandle

QueryMode

RaiseTpl

ResetSystem

RestoreTpl

SetMode

SetTime

SetVirtualAddressMap

SetWakeupTime

SetWatchdogTimer

Stall

Status

TextReset

TextString

Tpl

VirtualAddress